Update namespaces after move of Python utils to Common
