+++ qt/widgets/mplcpp/inc/MantidQtWidgets/MplCpp/Figure.h
@@ -47,9 +47,10 @@ public:
   void setFaceColor(const char *color);
   Axes addAxes(double left, double bottom, double width, double height);
   Axes addSubPlot(const int subplotspec, const QString projection = "");
-  Python::Object colorbar(const ScalarMappable &mappable, const Axes &cax,
-                          const Common::Python::Object &ticks = Python::Object(),
-                          const Common::Python::Object &format = Python::Object());
+  Common::Python::Object
+  colorbar(const ScalarMappable &mappable, const Axes &cax,
+           const Common::Python::Object &ticks = Common::Python::Object(),
+           const Common::Python::Object &format = Common::Python::Object());
 };
 
 } // namespace MplCpp
