+++ qt/widgets/plotting/src/Mpl/PreviewPlot.cpp
@@ -27,7 +27,7 @@ using MantidQt::Widgets::MplCpp::Figure;
 using MantidQt::Widgets::MplCpp::FigureCanvasQt;
 using MantidQt::Widgets::MplCpp::Line2D;
 using MantidQt::Widgets::MplCpp::MantidAxes;
-namespace Python = MantidQt::Widgets::MplCpp::Python;
+namespace Python = MantidQt::Widgets::Common::Python;
 
 namespace {
 Mantid::Kernel::Logger g_log("PreviewPlot");
@@ -134,7 +134,8 @@ void PreviewPlot::addSpectrum(const QString &lineName, const QString &wsName,
 void PreviewPlot::removeSpectrum(const QString &lineName) {
   auto axes = m_canvas->gca();
   Mantid::PythonInterface::GlobalInterpreterLock lock;
-  const auto lineNameAsUnicode = Python::NewRef(PyUnicode_FromString(lineName.toLatin1().constData()));
+  const auto lineNameAsUnicode =
+      Python::NewRef(PyUnicode_FromString(lineName.toLatin1().constData()));
   axes.forEachArtist("lines", [&lineNameAsUnicode](Artist &&line) {
     if (lineNameAsUnicode == line.pyobj().attr("get_label")())
       line.remove();
